fastapi/requirements.txt
fastapi[standard]==0.115.0
uvicorn==0.31.0
sqlalchemy==2.0.35
pymysql==1.1.1
python-dotenv==1.0.1
pydantic==2.9.2
pytest==8.3.3
pytest-cov==5.0.0
boto3==1.35.29
moto==5.0.16
google-cloud-storage===2.18.2